TIDs aren't all integers?

+++ dags/dataset_ETL_GEBCO_netcdf_TID_to_pgsql.py
@@ -52,7 +52,7 @@ def netcdf_to_points():
         CREATE TABLE gebco_tid (
             latitude DECIMAL,
             longitude DECIMAL,
-            tid INTEGER
+            tid NUMERIC
         );
     """
